[mob][photos] Cleanup ImageMlIsolate
